+++ lightly_studio/src/lightly_studio/resolvers/image_resolver/get_adjacent_images/__init__.py
@@ -1,0 +1,7 @@
+"""Resolver for the previous/next image of a sample in the current sort order."""
+
+from lightly_studio.resolvers.image_resolver.get_adjacent_images._dispatch import (
+    get_adjacent_images,
+)
+
+__all__ = ["get_adjacent_images"]
